Update Again Twenty Five!.cpp

+++ 800/Again Twenty Five!.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-	ll n;
+	int n;
 	cin>>n;
 	cout<<25<<endl;
 }
